inclusao do serviço de banco no docker

+++ src/main/java/com/amcom/order/dto/OrderItemDTO.java
@@ -3,9 +3,7 @@ package com.amcom.order.dto;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import lombok.Builder;
 
-import java.math.BigDecimal;
-
 @Builder
 @JsonSerialize
-public record OrderItemDTO(String sku, int quantity, BigDecimal unitPrice) {
+public record OrderItemDTO(String sku, Integer quantity, Double unitPrice) {
 }
